change to double

+++ src/lib/Math/Algebra.java
@@ -3,24 +3,24 @@ package lib.Math;
 import java.util.Arrays;
 
 public abstract class Algebra {
-    public static float[] linear2x2Equation(float a, float b, float c, float d, float e, float f) {
+    public static double[] linear2x2Equation(double a, double b, double c, double d, double e, double f) {
 
         //ax+by=e
         //cx+dy=f
-        float discriminant=(a*d)-(b*c);
+        double discriminant=(a*d)-(b*c);
 
         if (discriminant== 0)return null;
-        float x = ((e * d) - (b * f)) / ((a * d) - (b * c));
-        float y = ((a * f) - (e * c)) / ((a * d) - (b * c));
-        float[] z = {x, y};
+        double x = ((e * d) - (b * f)) / ((a * d) - (b * c));
+        double y = ((a * f) - (e * c)) / ((a * d) - (b * c));
+        double[] z = {x, y};
         return z;
     }
 
-    public static double[] quadraticEquation(float a, float b, float c){
+    public static double[] quadraticEquation(double a, double b, double c){
         //
         //
         double[] r;
-        float discriminant=(b*b)-(4f*a*c);
+        double discriminant=(b*b)-(4f*a*c);
         if (discriminant>0){
             r = new double[2];
             r[0]=(-b+Math.sqrt(discriminant))/(2f*a);
